requirements.txt
boto3>=1.34.0
faiss-cpu>=1.7.4
numpy>=1.24.0
streamlit>=1.28.0
langchain-text-splitters>=0.0.1
langchain-community>=0.0.1
PyPDF2>=3.0.0
